Clean up unused install files.
Eventually trying to cleanup the nvidia install where we are having an
issue with the cuda driver that randomly starts happening if docker is
not safely stopped. This commit is just doing a bit of cleanup before we
get to those real fixes in later commits.

Tested by:

1) Used packer to build ami:

    (local)$ packer build \
        -var 'aws_access_key=<ACCESS_KEY>' \
        -var 'aws_secret_key=<SECRET_KEY>' \
        gpu-packer.json

2) Created GPU instance (p2.xlarge) from AMI.

3) Cloned https://github.com/churchlab/mlpe-gfp-pilot.

4) Built GPU docker image:

    (aws)$ docker build -f docker/Dockerfile.gpu -t mlpe-gfp-pilot-gpu .

5) Ran a docker container from the image:

    (aws)$ nvidia-docker run -it -p 8888:8888 -p 6006:6006 -v ~/notebooks:/notebooks mlpe-gfp-pilot-gpu

Confirmed no error.